Moving Rational class to utility package
